GDPR-Event-handling: Refactored the code
