+++ izettle-gdpr/pom.xml
@@ -10,6 +10,12 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>izettle-gdpr</artifactId>
+
+    <properties>
+        <izettle.messaging.version>1.0.113</izettle.messaging.version>
+        <aws.java.sdk.version>1.6.4</aws.java.sdk.version>
+    </properties>
+
     <dependencies>
         <dependency>
             <groupId>com.amazonaws</groupId>
@@ -30,7 +36,7 @@
         <dependency>
             <groupId>com.izettle.toolbox</groupId>
             <artifactId>izettle-messaging</artifactId>
-            <version>1.0.113</version>
+            <version>${izettle.messaging.version}</version>
         </dependency>
         <dependency>
             <groupId>io.dropwizard</groupId>
@@ -40,7 +46,7 @@
         <dependency>
             <groupId>com.amazonaws</groupId>
             <artifactId>aws-java-sdk</artifactId>
-            <version>1.6.4</version>
+            <version>${aws.java.sdk.version}</version>
         </dependency>
     </dependencies>
