+++ izettle-gdpr/src/main/java/com/izettle/gdpr/dao/GdprDao.java
@@ -1,0 +1,15 @@
+package com.izettle.gdpr.dao;
+
+import com.izettle.gdpr.cassandra.Paging;
+import com.izettle.gdpr.cassandra.PagingResult;
+import com.izettle.gdpr.model.GdprEvent;
+
+public interface GdprDao {
+
+    void deleteAll(GdprEvent event);
+
+    PagingResult<?> getAll(Paging paging);
+
+    String getDescription();
+
+}
