+++ izettle-gdpr/src/main/java/com/izettle/gdpr/messaging/GdprEventMessageHandler.java
@@ -1,8 +1,9 @@
+package com.izettle.gdpr.messaging;
+
+import com.izettle.gdpr.model.GdprEvent;
 import com.izettle.messaging.handler.MessageHandler;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
-import model.GdprEvent;
-import model.GdprEventHandler;
 
 @Data
 @Slf4j
@@ -12,7 +13,6 @@ public class GdprEventMessageHandler implements MessageHandler<GdprEvent> {
 
     @Override
     public void handle(GdprEvent gdprEvent) throws Exception {
-
         switch (gdprEvent.getType()) {
             case FORGET:
                 handler.handleForget(gdprEvent);
@@ -23,6 +23,5 @@ public class GdprEventMessageHandler implements MessageHandler<GdprEvent> {
             default:
                 break;
         }
-
     }
 }
