+++ izettle-gdpr/src/main/java/com/izettle/gdpr/model/GdprEvent.java
@@ -1,4 +1,4 @@
-package model;
+package com.izettle.gdpr.model;
 
 import java.util.UUID;
 import lombok.Data;
