+++ izettle-gdpr/src/main/java/com/izettle/gdpr/model/GdprStatusReportMessage.java
@@ -1,4 +1,4 @@
-package model;
+package com.izettle.gdpr.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
