--- izettle-gdpr/src/main/java/model/GdprDao.java
@@ -1,13 +1,0 @@
-package model;
-
-import java.util.UUID;
-
-public interface GdprDao {
-
-    void deleteAll(GdprEvent event);
-
-    PagingResult<?> getAll(Paging paging);
-
-    String getDescription();
-
-}
